# apps/openapi.yaml
openapi: 3.0.0
paths:
...
  /trust/verify/{receiptId}:
    get:
      operationId: TrustMongoController_verifyReceipt
      summary: Verify a ballot receipt SHA-256 against the MongoDB Merkle root
      parameters:
        - name: receiptId
          required: true
          in: path
          description: Ballot ID or hash
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
...
  /api/v1/events/{eventId}/ranking/preview:
    get:
      operationId: RankingController_previewRankings
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/events/{eventId}/ranking/simulate-weights:
    post:
      operationId: RankingController_simulateWeights
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
  /api/v1/events/{eventId}/ranking/disagreement/trigger:
    post:
      operationId: RankingController_triggerDisagreementReviews
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
  /api/v1/events/{eventId}/ranking/publish:
    post:
      operationId: RankingController_publishResults
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
  /api/v1/events/{eventId}/ranking/published:
    get:
      operationId: RankingController_getPublishedResults
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/events/{eventId}/ranking/pairwise:
    get:
      operationId: RankingController_getPairwiseRankings
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/events/{eventId}/ranking/normalization-proof:
    get:
      operationId: RankingController_getNormalizationProof
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/events/{eventId}/ranking/{runId}/sign-off:
    post:
      operationId: RankingController_signOff
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: runId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
  /api/v1/events/{eventId}/ranking/projects/{projectId}/feedback:
    get:
      operationId: RankingController_getEntrantFeedback
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: projectId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/events/{eventId}/idea-reports:
    post:
      operationId: IntelligenceController_createReport
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
  /api/v1/events/{eventId}/idea-reports/{id}:
    get:
      operationId: IntelligenceController_getReport
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
    delete:
      operationId: IntelligenceController_deleteReport
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/trust/verify/{eventId}:
    get:
      operationId: TrustController_verifyChain
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
...
  /api/v1/trust/events/{eventId}/audit-trail:
    get:
      operationId: TrustController_getAuditTrail
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
  /api/v1/trust/events/{eventId}/export:
    get:
      operationId: TrustController_getExport
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
...
info:
  title: DOGFOOD OS API
  description: Self-hosted, offline-first operating system for fair judged events
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...